Week #2 - Day #5 - Hackathon/Daily Challenges/Restaurant Menu Manager with Flask/requirements.txt
Flask==3.0.3
psycopg[binary]==3.2.10
python-dotenv==1.0.1
